apps/node-agent/pyproject.toml
[project]
name = "node-agent"
version = "0.1.0"
description = "Node agent for Pi"
authors = [
    {name = "mdbro",email = "56290383+mdbro@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.120.0,<0.121.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "aiomqtt (>=2.0.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "psutil (>=6.0.0,<7.0.0)",
    "cryptography (>=42.0.0,<43.0.0)",
    "zeroconf (>=0.132.2,<0.133.0)",
    "dbus-next (>=0.2.3,<0.3.0)",
    "bleak (>=0.22.3,<0.23.0) ; python_version < '3.14'",
    "gpiozero (>=2.0.1,<3.0) ; platform_system == 'Linux' and (platform_machine == 'aarch64' or platform_machine == 'armv7l')",
    "lgpio (>=0.2.2,<1.0) ; platform_system == 'Linux' and (platform_machine == 'aarch64' or platform_machine == 'armv7l')",
    "spidev (>=3.6,<4.0) ; platform_system == 'Linux' and (platform_machine == 'aarch64' or platform_machine == 'armv7l')",
    "pigpio (>=1.78,<2.0) ; platform_system == 'Linux' and (platform_machine == 'aarch64' or platform_machine == 'armv7l')",
    "opentelemetry-api (>=1.27.0,<2.0.0)",
    "opentelemetry-sdk (>=1.27.0,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.27.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.48b0,<0.49)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
